requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
pinecone-client>=3.0.0
PyPDF2>=3.0.0
streamlit>=1.32.0 